Update footer
Just figured it wasn't reason enough to click on the link in the footer about our analytics being more accurate than ga4. Stats also show that clicks on this from the footer are lower as compared to other links in the same category. 

Feel free to reject the PR if it doesn't resonate.